requirements.txt
transformers==4.37.2
torch==2.1.2
accelerate==0.27.2
python-telegram-bot==20.7
sentencepiece==0.2.0
protobuf==3.20.3